# docs/swagger.yaml
definitions:
  domain.Metric:
    properties:
      delta:
        description: значение метрики в случае передачи counter
        type: integer
      id:
        description: имя метрики
        type: string
      type:
        description: параметр, принимающий значение gauge или counter
        type: string
      value:
        description: значение метрики в случае передачи gauge
        type: number
    type: object
info:
  contact: {}
paths:
  /update:
    post:
      consumes:
      - application/json
      description: Обновляет метрику на основе JSON тела запроса
      parameters:
      - description: Данные метрики
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Metric'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная метрика
          schema:
            $ref: '#/definitions/domain.Metric'
        "400":
          description: Ошибка в запросе
          schema:
            type: string
      tags:
      - Update
  /update/{type}/{name}/{value}:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Обновляет метрику по данным формы
      parameters:
      - description: Тип метрики
        in: formData
        name: type
        required: true
        type: string
      - description: Название метрики
        in: formData
        name: name
        required: true
        type: string
      - description: Значение метрики
        in: formData
        name: value
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      tags:
      - Update
  /updates:
    post:
      consumes:
      - application/json
      description: Обновляет массив метрик на основе JSON тела запроса
      parameters:
      - description: Массив данных метрик
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/domain.Metric'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Массив обновленных метрик
          schema:
            items:
              items:
                $ref: '#/definitions/domain.Metric'
              type: array
            type: array
        "400":
          description: Ошибка в запросе
          schema:
            type: string
      tags:
      - Update
  /value:
    post:
      consumes:
      - application/json
      description: Принимает метрику без значения и возвращает тот же объект метрики
        с значением
      parameters:
      - description: Данные метрики
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Metric'
      produces:
      - application/json
      responses:
        "200":
          description: Метрика с значением
          schema:
            $ref: '#/definitions/domain.Metric'
        "400":
          description: Ошибка в запросе
          schema:
            type: string
      tags:
      - Update
  /value/{type}/{name}:
    get:
      consumes:
      - application/json
      description: Получает значение метрики на основе типа и имени
      parameters:
      - description: Тип метрики
        in: path
        name: type
        required: true
        type: string
      - description: Название метрики
        in: path
        name: name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Значение метрики
          schema:
            type: string
        "400":
          description: Ошибка в запросе
          schema:
            type: string
      tags:
      - Update
swagger: "2.0"